refactor: extract duplicated insertTemplate logic into useInsertTemplate hook
